+++ packages/tests/package.json
@@ -50,7 +50,6 @@
     "react-dom": "^18.2.0",
     "runtypes": "^6.6.0",
     "scale-codec": "^0.13.0",
-    "superjson": "^1.12.4",
     "superstruct": "^1.0.0",
     "tsx": "^4.0.0",
     "tupleson": "0.23.1",
@@ -61,7 +60,6 @@
     "vitest-environment-miniflare": "^2.12.0",
     "ws": "^8.0.0",
     "yup": "^1.0.0",
-    "zod": "^3.0.0",
     "zod-form-data": "^2.0.1"
   }
 }
